requirements.txt
Django==4.2.7
django-model-utils==4.3.1
django-notifications==0.1.dev0
django-notifications-hq==1.8.3
jsonfield==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
swapper==1.3.0
typing_extensions==4.8.0
gunicorn==20.1.0